Renaming library to stn and restructuring files Adding module for methods to solve stp problems
