+++ stn/methods/dsc_lp.py
@@ -28,7 +28,7 @@ import sys
 import logging
 from allocation.utils.config_logger import config_logger
 from math import ceil
-import os
+from pathlib import Path
 
 """
 Computes the Degree of Strong Controllability (DSC) using an LP program as presented in:
@@ -41,10 +41,9 @@ MAX_FLOAT = sys.float_info.max
 
 
 class DSC_LP(object):
-    code_dir = os.path.abspath(os.path.dirname(__file__))
-    main_dir = os.path.dirname(code_dir)
-    config_logger(main_dir + '/config/logging.yaml')
-    logger = logging.getLogger('stp.dsc_lp')
+    p = Path(__file__).parents[2]
+    config_logger(str(p) + '/config/logging.yaml')
+    logger = logging.getLogger('stn.dsc_lp')
 
     def __init__(self, stnu):
         self.stnu = stnu
