--- stp/temporal_networks/pstn/__init__.py
@@ -1,6 +1,0 @@
-__title__ = 'pstn'
-__version__ = '0.1.0'
-
-# from .node import NodePSTN
-from .constraint import Constraint
-from .pstn import PSTN
