--- stp/temporal_networks/stn/__init__.py
@@ -1,5 +1,0 @@
-__title__ = 'stn'
-__version__ = '0.1.0'
-
-from .node import Node
-from .stn import STN
