--- stp/temporal_networks/stnu/__init__.py
@@ -1,4 +1,0 @@
-__title__ = 'stnu'
-__version__ = '0.1.0'
-
-from .stnu import STNU
