+++ test/update_stnu.py
@@ -1,4 +1,4 @@
-from stp.temporal_networks.stnu import STNU
+from stn.stnu.stnu import STNU
 import unittest
 import logging
 import sys
@@ -12,7 +12,7 @@ logger.addHandler(stream_handler)
 
 
 class UpdateSTNU(unittest.TestCase):
-    logger = logging.getLogger('stp.test')
+    logger = logging.getLogger('stn.test')
 
     def setUp(self):
         self.tasks = load_dataset('three_tasks.csv')
